fix: ProtectedRoute
